Moved code to generate subsystem test into separate function
